fix: isOK를 int형으로 변환하고 테스트

+++ Kirby_SupernovaVanguard/Kirby_SupernovaVanguard/SelectScene.cpp
@@ -59,7 +59,7 @@ SelectScene::~SelectScene()
 
 void SelectScene::DrawBitmapDoubleBuffering(HWND hWnd, HDC hdc, RECT& rectView, vector<Object*>& clients, POINT curMousePos, TOTALDATA& tData)
 {
-	if ((int)(tData.publicdata.isOK) != 0 && (int)(tData.publicdata.isOK) != 1)
+	if (tData.publicdata.isOK != 0 && tData.publicdata.isOK != 1)
 		return;
 
 	HDC hDoubleBufferDC;
